feat : mypage 중복 페이지 수정
